# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mousewheeld",
    "description": "The locomotion input of a braemons rig: a running wheel, read and published.\n\nUnits are in the names \u2014 `position_cm`, `counts_per_cm`, `velocity_cm_s`. Counts are the device's only unit; everything that leaves here is centimetres. Unknown fields are refused rather than defaulted.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/calibration": {
      "get": {
        "tags": [
          "calibration"
        ],
        "operationId": "read_calibration",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calibration"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "calibration"
        ],
        "summary": "Change a calibration by hand. Every field optional; the rest is left alone.",
        "operationId": "replace_calibration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calibration"
                }
              }
            }
          },
          "404": {
            "description": "no such axis"
          }
        }
      }
    },
    "/api/calibration/ball": {
      "get": {
        "tags": [
          "calibration"
        ],
        "summary": "The 2-D skeleton: modelled, routed, and refused by name.",
        "operationId": "read_ball",
        "responses": {
          "501": {
            "description": "arrives with the hardware"
          }
        }
      },
      "put": {
        "tags": [
          "calibration"
        ],
        "operationId": "replace_ball",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BallCalibration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "arrives with the hardware"
          }
        }
      }
    },
    "/api/calibration/measure/apply": {
      "post": {
        "tags": [
          "calibration"
        ],
        "summary": "Make the measurement the rig's truth.",
        "operationId": "apply_measurement",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementApplied"
                }
              }
            }
          },
          "409": {
            "description": "nothing measured"
          }
        }
      }
    },
    "/api/calibration/measure/finish": {
      "post": {
        "tags": [
          "calibration"
        ],
        "summary": "Stop counting and report \u2014 measured beside configured, and what the\ndifference looks like.",
        "operationId": "finish_measurement",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementResult"
                }
              }
            }
          },
          "409": {
            "description": "nothing is being measured"
          }
        }
      }
    },
    "/api/calibration/measure/start": {
      "post": {
        "tags": [
          "calibration"
        ],
        "summary": "Note the counter and start counting.",
        "operationId": "start_measurement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMeasurement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementStarted"
                }
              }
            }
          },
          "503": {
            "description": "no board"
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "config"
        ],
        "operationId": "read_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "config"
        ],
        "operationId": "patch_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigView"
                }
              }
            }
          },
          "422": {
            "description": "refused"
          }
        }
      }
    },
...
    "/api/device/connect": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Open the link, or say why not.",
        "operationId": "connect",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceInfo"
                }
              }
            }
          },
          "503": {
            "description": "no board on this host"
          }
        }
      }
    },
...
    "/api/position/zero": {
      "post": {
        "tags": [
          "state"
        ],
        "summary": "Move the API origin \u2014 never the accumulator a camera differences.",
        "operationId": "zero_position",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZeroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RigState"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/zone-sets/{name}": {
      "get": {
        "tags": [
          "zones"
        ],
        "operationId": "read_zone_set",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "the set's name in the store",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneSet"
                }
              }
            }
          },
          "404": {
            "description": "no such set"
          }
        }
      },
      "put": {
        "tags": [
          "zones"
        ],
        "summary": "Write a set to the store. Stored as authored \u2014 `$name` references intact,\ncentimetres intact \u2014 because the store is what a person edits and copies\nbetween rigs. Compilation happens at arm.",
        "operationId": "replace_zone_set",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "the set's name in the store",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneSet"
                }
              }
            }
          },
          "422": {
            "description": "refused"
          }
        }
      }
    },
    "/api/zone-sets/{name}/validate": {
      "post": {
        "tags": [
          "zones"
        ],
        "summary": "Compile against the current calibration and the board's capacities, and\nupload nothing.",
        "description": "This is what lets a zone set be wrong in an editor rather than at arm time,\nwhich is the difference between a message on screen and a TTL that fires in\nthe wrong place in the middle of a session.",
        "operationId": "validate_zone_set",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "the set's name in the store",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/zones/arm": {
      "post": {
        "tags": [
          "zones"
        ],
        "summary": "Resolve the patch, compile, upload, wait for `armed`.",
        "operationId": "arm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmedZones"
                }
              }
            }
          },
          "404": {
            "description": "no set by that name"
          },
          "422": {
            "description": "the set did not compile"
          },
          "503": {
            "description": "no board"
          }
        }
      }
    },
...
    "/api/zones/save": {
      "post": {
        "tags": [
          "zones"
        ],
        "summary": "Write the armed set, and its armed state, to the board's flash.",
        "operationId": "save_to_flash",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceInfo"
                }
              }
            }
          },
          "409": {
            "description": "nothing armed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ArmRequest": {
        "type": "object",
        "description": "`POST /api/zones/arm`.",
        "required": [
          "zone_set"
        ],
        "properties": {
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque text stored with the arm \u2014 triald writes `trial 42`. The daemon\nstays trial-blind and hands it back unread."
          },
          "origin": {
            "$ref": "#/components/schemas/ArmOrigin"
          },
          "patch": {
            "type": "object",
            "description": "Per-trial values for this set's `$name` bounds.",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "zone_set": {
            "type": "string",
            "description": "A name from the store. Never a slot: the daemon holds the sets, and a\nname it does not have is a configuration error the caller can report."
          }
        },
        "additionalProperties": false
      },
...
      "BallCalibration": {
        "type": "object",
        "description": "Declared, refused, and here so the shape of the answer is not a surprise.",
        "required": [
          "diameter_cm",
          "sensor_angles_deg"
        ],
        "properties": {
          "diameter_cm": {
            "type": "number",
            "format": "double"
          },
          "sensor_angles_deg": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "description": "Where each sensor sits, in degrees around the ball."
          }
        },
        "additionalProperties": false
      },
      "Calibration": {
        "type": "object",
        "description": "Every axis' calibration, plus the 2-D skeleton.",
        "required": [
          "axes"
        ],
        "properties": {
          "axes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AxisCalibration"
            }
          },
          "ball": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BallCalibration",
                "description": "The ball's transform \u2014 sensor mounting angle, ball diameter, the map\nonto `x`, `y` and `yaw`. Always `None` today; its routes answer 501. It\nexists in the model so that adding the hardware is filling in rather\nthan redesigning."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "CalibrationPatch": {
        "type": "object",
        "required": [
          "axes"
        ],
        "properties": {
          "axes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AxisCalibrationPatch"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ConfigPatch": {
        "type": "object",
        "description": "`PATCH /api/config`. Absent fields are left alone.",
        "properties": {
          "display_hz": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "rate_hz": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "ring_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "ConfigView": {
        "type": "object",
        "description": "`GET /api/config` \u2014 the settings a session may change, and the ones it may\nonly read, in one answer.",
        "required": [
          "rate_hz",
          "display_hz",
          "ring_minutes",
          "shm_name",
          "shm_open",
          "shm_writes",
          "event_port",
          "port",
          "starves_the_display"
        ],
        "properties": {
          "display_hz": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "event_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "port": {
            "type": "string",
            "description": "Read-only here: wiring is changed where wiring is described."
          },
          "rate_hz": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "ring_minutes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "shm_name": {
            "type": "string"
          },
          "shm_open": {
            "type": "boolean",
            "description": "Whether that segment is actually open. A name in a config file and a\nmapped segment are different things, and the difference is the whole\nreason a corridor sometimes does not move."
          },
          "shm_writes": {
            "type": "integer",
            "format": "int64",
            "description": "Writes that have reached it since it was created.",
            "minimum": 0
          },
          "starves_the_display": {
            "type": "boolean",
            "description": "True when `rate_hz` is at or below `display_hz`."
          }
        }
      },
      "DeviceInfo": {
        "type": "object",
        "required": [
          "connected",
          "port",
          "board",
          "firmware_version",
          "protocol_version",
          "uptime_device_us",
          "capacities",
          "link"
        ],
        "properties": {
          "board": {
            "type": "string",
            "description": "`teensy41`, `esp32`, or `simulated`."
          },
          "capacities": {
            "$ref": "#/components/schemas/Capacities"
          },
          "connected": {
            "type": "boolean"
          },
          "firmware_version": {
            "type": "string"
          },
          "flashed_zone_set": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FlashedZoneSet",
                "description": "The set written to the board's flash, if any: a standalone rig comes up\nwith its zones, and the board reports which copy it holds."
              }
            ]
          },
          "link": {
            "$ref": "#/components/schemas/LinkStats"
          },
          "port": {
            "type": "string",
            "description": "The serial port this daemon owns, from the rig config."
          },
          "protocol_version": {
            "type": "integer",
            "format": "int32",
            "description": "The wire protocol version the board greeted with.",
            "minimum": 0
          },
          "uptime_device_us": {
            "type": "integer",
            "format": "int64",
            "description": "The board's clock since it booted \u2014 a reset shows up here before it\nshows up anywhere else.",
            "minimum": 0
          }
        }
      },
...
      "MeasurementApplied": {
        "type": "object",
        "required": [
          "axis",
          "counts_per_cm",
          "zone_sets_invalidated"
        ],
        "properties": {
          "axis": {
            "type": "string"
          },
          "counts_per_cm": {
            "type": "number",
            "format": "double"
          },
          "zone_sets_invalidated": {
            "type": "boolean",
            "description": "Compiled zone sets are stale now, and the next arm recompiles. A set\narmed under one calibration is never silently reinterpreted under\nanother."
          }
        }
      },
      "MeasurementResult": {
        "type": "object",
        "description": "What the wheel said, beside what the config says. Never applied on its own:\na calibration change invalidates every compiled zone set, so it is a second,\ndeliberate press.",
        "required": [
          "axis",
          "known_distance_cm",
          "counts",
          "measured_counts_per_cm",
          "configured_counts_per_cm"
        ],
        "properties": {
          "axis": {
            "type": "string"
          },
          "configured_counts_per_cm": {
            "type": "number",
            "format": "double"
          },
          "counts": {
            "type": "integer",
            "format": "int64"
          },
          "counts_per_rev": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "decoding_suspect": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set when the measurement is a whole-number multiple of the configured\nvalue: a decoding mistake rather than a wheel, and worth saying in\nwords before somebody applies it."
          },
          "implied_circumference_cm": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "The circumference the measurement implies, when the counts per\nrevolution are known \u2014 the number to hold against the wheel with a tape."
          },
          "known_distance_cm": {
            "type": "number",
            "format": "double"
          },
          "measured_counts_per_cm": {
            "type": "number",
            "format": "double"
          },
          "nominal_counts_per_cm": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "What the wheel's stated dimensions imply \u2014 the arithmetic this\nmeasurement exists to correct. On the report rather than left to a\nclient, so a result is readable on its own a month later."
          }
        }
      },
      "MeasurementStarted": {
        "type": "object",
        "description": "A measurement in progress.",
        "required": [
          "axis",
          "known_distance_cm",
          "counts_at_start",
          "counts"
        ],
        "properties": {
          "axis": {
            "type": "string"
          },
          "counts": {
            "type": "integer",
            "format": "int64"
          },
          "counts_at_start": {
            "type": "integer",
            "format": "int64",
            "description": "The counter when counting began, so a UI can show progress without\nasking the daemon to keep a second copy of it."
          },
          "known_distance_cm": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "RigState": {
        "type": "object",
        "required": [
          "axes",
          "health",
          "link"
        ],
        "properties": {
          "axes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AxisState"
            }
          },
          "health": {
            "$ref": "#/components/schemas/LinkHealth"
          },
          "link": {
            "$ref": "#/components/schemas/LinkState"
          }
        }
      },
...
      "StartMeasurement": {
        "type": "object",
        "description": "`POST /api/calibration/measure/start`.",
        "required": [
          "axis",
          "known_distance_cm"
        ],
        "properties": {
          "axis": {
            "type": "string"
          },
          "known_distance_cm": {
            "type": "number",
            "format": "double",
            "description": "How far the wheel is about to be rolled, by hand, along its running\nsurface. A metre in one continuous motion: a metre averages away where\nyou started and stopped, and one revolution does not."
          }
        },
        "additionalProperties": false
      },
...
      "ValidationReport": {
        "type": "object",
        "description": "`POST /api/zone-sets/{name}/validate` \u2014 compile without uploading.",
        "required": [
          "ok",
          "zone_count",
          "counts_per_cm",
          "references"
        ],
        "properties": {
          "counts_per_cm": {
            "type": "number",
            "format": "double",
            "description": "The calibration it was compiled against, so a report is readable a day\nlater without guessing which one was in force."
          },
          "ok": {
            "type": "boolean"
          },
          "problem": {
            "type": [
              "string",
              "null"
            ],
            "description": "The first thing wrong, in the words the compiler would use. `None` when\nit compiles."
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The `$name`s this set still needs at arm time."
          },
          "zone_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ZeroRequest": {
        "type": "object",
        "description": "`POST /api/position/zero`.\n\nMoves the **API origin** \u2014 what displacement, distance and the zones are\nmeasured from. Never the accumulator published to a camera: vstimd\ndifferences that every frame, and a value that jumps backwards jumps the\ncorridor backwards. Teleporting a corridor is a command to vstimd, from\nwhoever runs the experiment.",
        "properties": {
          "axes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Which axes. Empty means all of them."
          }
        },
        "additionalProperties": false
      },
...
      "ZoneSet": {
        "type": "object",
        "description": "A named set, as it is stored and as it is uploaded.\n\nDeclaration order resolves two zones firing in one scan: both still fire,\nand the order is the order of their `zone_hit` messages.",
        "required": [
          "zone_set_version",
          "zones"
        ],
        "properties": {
          "zone_set_version": {
            "type": "integer",
            "format": "int32",
            "description": "Bumped by whoever edits the set. The board reports the version it holds,\nso \"is the board running what I think it is\" is one comparison.",
            "minimum": 0
          },
          "zones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Zone"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}